requirements.txt
click==8.1.3
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymemcache==4.0.0
PyYAML==6.0
slack-sdk==3.19.4
Werkzeug==2.2.2
zipp==3.10.0
